敵とステージの判定使いまわせるように

+++ EnemyBase.cpp
@@ -77,7 +77,7 @@ int EnemyBase::GetHitFrameCnt()
 	return hitFrameCounter;
 }
 
-int EnemyBase::GetDamage() {
+float EnemyBase::GetDamage() {
 	return damage;
 }
